Factor out ledgerSpec function.
