+++ lib/fine-types/test/data/Cardano/Ledger/Shelley/Tx.fine
@@ -38,7 +38,7 @@ TxBody =
     , txttl    : Slot  -- time to live
     , txwdrls  : Wdrl  -- withdrawals
     , txup     : Update?  -- protocol parameter update
-    , txMDhash : MetadataHash?  -- metadata hash  
+    , txMDhash : MetadataHash?  -- metadata hash
     };
 
 TxWitness =
